requirements.txt
Flask
langchain
langchain-community
langchain-core
google-generativeai
chromadb
pandas
langchain-google-genai
python-dotenv